Cluster scoped Bus (#117)
* Cluster scoped Bus

Make the Bus resource cluster scpoped instead of namespaced. Bus
releated deployments will be put into the knative-eventing namespace
and use a shared serviceaccount for all bus provisioners and
dispatchers.

This PR does not change the name of the Bus resource to ClusterBus.
That should happen later if we desire to have both cluster wide and
namespaced buses.

Subscribers using a short DNS name will automatically be expanded to
use the full name.namespace.svc.cluster.local form based on the
namespace of the subscription resource. Other bus implementations will
also need to make this change.

Fixes #106

* Rename Bus to ClusterBus

* Restore namespaced Bus support

Channels can target either a Bus or a ClusterBus, but never both.

* doc polish

* Enforce Bus and ClusterBus are not both used in a Channel

* review feedback

* Channel CRD Status properties are no longer pointers