Remove feed finalizer on job failure (#255)
* Record events on start and stop job transitions

The feed controller now records events when jobs are created, complete,
or fail. This should make debugging feed issues slightly easier.

* Use the FeedSuccess and FeedFailed reasons

Conditions when stopping should use the same reasons as conditions when
starting, at least for now.

* Remove finalizer when stop job fails

Without this it's impossible to delete a malformed feed without manually
editing it.

Debugging a feed that fails to be stopped is slightly more possible now
that events are recorded on job failure (events continue to exist after
their original object is deleted).