Advanced Trigger Filtering Design Proposal (#1047)
* Add a proposal for advanced filtering

* fix typo in example

* Add caveat about variable prefixes

* pluralize CloudEvents

* fix formatting formatted by formatter

* Improve readability of complex expression example

* CloudEvents fields are called attributes

* Fix table formatting

(From Ahmed)

* Move extensions to the `ce` prefix

This makes extension attributes easier to use and promote to official
attributes. The `ext` prefix has been removed.

* Expand prefix match example

Mention startsWith, match, and matches functions.

* Check expression syntax in the Trigger webhook

* Add versioning caveat

* Reimplement SourceAndType as an expression

* Remove parse flags

Extensions are always parsed anyway by the CloudEvents SDK, so setting
parseExtensions to false has no efficiency benefit. Removing the field
improves the approachability of the filtering interface.

We can detect when parsing the expression whether data fields are used.
This provides a more reliable signal than the parseData flag, so the
flag is unnecessary.

* Simplify syntax and add attributes filter

The `cel` level of the expression filter is removed and `expression` is
moved up to its place.

An `attributes` filter is added to replace `sourceAndType` (now
deprecated).

Filter combination at the top level is explicitly disallowed.

Examples have been added to clarify some unusual cases.

* Note that CloudEvents is vague and in flux

CloudEvents is still figuring out whether it wants to support numeric
values in extensions. Note that there will probably be dragons for a
while.

* Clarify structured filters alternative

This "alternative considered" is about using structured filters
exclusively, without an expression language. Be clear that structured
filters are fine as long as they compile down to an expression.

* Explain choice of exclusive top-level filter

* Add filter spec alternatives considered

* Correct inequality example

* Add open questions section

Add open questions related to evaluation errors.