Trigger filter metrics (#1177)
* Export metrics from the broker filter pod

Same setup as the broker ingress pod: OpenCensus Prometheus exporter
listening on :9090.

*  Instrument trigger receiver

The receiver now records an overall event received count per trigger,
with accept/reject result tags, and a filtering time histogram with
pass/fail/error result tags.

* Rename and document Broker and Trigger metrics

Rename existing Broker metrics with a more concise namespace and
replace `messages` with `events` in metric names and descriptions.

Add documentation for Trigger metrics.

These metrics are now exported by the Broker and Trigger data plane
pods:

* broker_event_total
* broker_dispatch_time
* trigger_events_total
* trigger_filter_time

* Add metrics port to Pod and Service specs

Also added port 8080 for the filter pod's CloudEvents transport.

* Remove unused TagError

Also added TagResult to the filter time view.

* Hardcode namespace in the view

This eliminates the possibility of view name ambiguity and makes clearer
in the view definition what the metric will be called. Metrics namespace
is a Prometheus-specific option so it may not apply to all exporters.

Specifically this is required to have a `dispatch_time` metric for
the broker receiver.

* Add trigger dispatch time metric

This is the timing of actually handling the event, including
subscriber latency.

* Update test fixtures for new port advertisements

* Change dispatch time buckets for greater precision

Since Broker ingress dispatches to channel(s), this should normally be
fast.

* Switch to 1,2,5,10... buckets

Added a Buckets125 method to generate these sequences for calling
view.Distribution.

* Rename reject tag value to drop

Drop is consistent with iptables definitions: REJECT notifies the
sender, DROP doesn't. In this case, the sender gets a 202 back even if
the event doesn't pass the filter.

* Remove code missed in merge

* Add container ports to Broker expected creates

* Tag metrics with the result of the send

If the CE client returns an error, that should be tagged with
result=error instead of result=accept.

* Remove obsolete comment

The broker receiver initiates graceful shutdown by canceling the context
passed to the CE client when the stop channel is closed.