Broker latency test converted to use Mako (#1728)
* Convert broker latency test to a Mako test

Mostly a direct port with Mako bits sprinkled in. Also some yaml files
for running in an existing cluster.

* Remove 30 second sleep on start

This was in place for the previous test which started everything at once.
Now we can assume that the test infrastructure is already in place.

* Use sendTime for all sample points

Correlates the publish and delivery latency of each event by time on
the graph.

* Add aggregate publish and deliver error counts

* Use shared Mako config code

Requires the mako-config ConfigMap in the perf-eventing namespace.

* Add required kodata files

* Remove double copyright notice

* Use mako.XTime to generate timestamp floats

* Revert accidental removal of ReplaceAll

* Remove the TestGrid support code

* Regen Gopkg.lock