Stop using PrependGenerateNameReactor. (#1895)
This is broken by the K8s test libs starting in 1.14, and to work properly we would need to patch vendor to partially revert the library change.

The intent of this is to remove the logic to avoid having someone unknowingly start to depend on this functionality, which keeps eventing from having to carry this patch.