Update Mako configs and add samples (#2215)
* Update Mako configs and add samples

Clarifies how to create a new benchmark config and the difference
between the test/test_images/performance config and the configs in
test/performance/benchmarks.

* Update config-mako configmap in CI perf namespaces

* Apply configmap instead of create

It may already exist

* Add quotes

* Fix whitespace and correct sample-prod.config

* README updates

* Remove configmap from benchmark setups

Instead, use the mako files in the directory as the source of truth and
create the configmap using kubectl create configmap --from-file as
documented in the README.

* Update test/performance/performance-tests.sh

Co-Authored-By: Chi Zhang <chizhg@google.com>

* Correct configmap creation steps

kubectl apply doesn't work with --from-file, so first delete then
create.