Create service accounts strictly after roles. (#2574)
* Create the `source-observer` role before a binding.

This is necessary when the installer of the eventing
YAML does not have `bind` on all roles in the cluster.
I've been prototyping an RBAC restriction that would
grant `escalate` on all roles but not `bind` on all
roles in the Operator RBAC. That adds some guard-rails
that are currently missing, but requires this particular
ordering to be correct.

* Move service accounts after all roles.