Feature: upgrade & downgrade tests (#2388)
* Change scripts to enable upgrade tests

* Upgrade tests overall structure

* Adding Istio & Serving to upgrade tests

* Delete non required Istio files

* Move prober to separate package

Prober will be used to it's own package to enable running it from within
e2e tests and upgrade tests.

* Working deploy & remove of configuration

* Probe receiver deploy & remove

Implemented receiver deployment and removal with getting service
node port, to later be able to connect and fetch results.

* Probe forwarder deploy & removal

* Switch to zap.SugarLogger for better messages visibility

* Inteligent waiting for all conditions

 * ksvc ready
 * triggers ready
 * pod ready
 * ksvc scales to zero

* Fix interval configuration for Sender

* Deploy & removal of sender pod

* Update ready message a bit

* Fetch receiver report via HTTP Get

* Don't deploy IMC from HEAD before each test

* Describe an upgrade tests

* Ensure knative-networking is installed just once

* Change error message to be more simplistic

Plus other minor review fixes

* A proper way of figure out a cluster hostname

* Better way of getting replicas while scale to zero

* Removing .editorconfig as not revelant to knative/eventing#2388

* Inteligent way of choosing serving version

With the script added serving version is choosen dynamilly, versus the
eventing version. The script tries to choose most compatibile version.

* Moving shellchecks to other PR

* Use Go's sync/errgroup for wait a bunch of conditions

* Use vendor for ensure.NoError

* Move awiat parameters to await package

* Remove trace span metric

* Use WaitForPodRunning from knative/pkg lib

* Use ConfigMap instead of Secret

* Warn about errors in upgrade tests (ref #2665)

* Move knative serving to duck & resources libs

The prober had to much code in itself. Moving knative serving to
more appropriate packages like duck and resources.

* Moving group await to lib package

* Create triggers with resource builders

* Add comment about wathola version

* Warn with FAIL to trigger reporting on Prow

* Fixing issues with scripts

* Move nodes resolver code in preparation for native/pkg#1001 merger

* Add notice about wathola donation

* Donation of wathola testing tool

The wathola testing tool was develop to make testing of eventing upgrade
tests possible. At first it was developed separately, and published with
versioning on https://quay.io/repository/cardil/wathola-*. It was done
that way to ease of use of the eventing prober in different context,
like eventing operator or knative vendor tests.

Then it was propose to donate that tool to knative. In
knative/test-infra#1631 discussion it was decided to donate it directly
to knative/eventing.

* Use test-infra scripts to intall serving and monitoring

* Removal of Istio & Serving from tests

* Wathola toml configuration respects default

After merging Wathola to eventing, a bug came out that toml package
has in 1.7.0 - https://github.com/pelletier/go-toml/issues/406

Adding tests to address that and pining pelletier/go-toml version to
working one.

* Updating codegen

* Cleanups after automatic review

* Fixing race conditions on wathola

* E2E Upgrade tests scripts

* Wathola receiver non-blocking and without race

* Migrate to cloudevents v2

* Move prober + wathola tool to upgrade directory

* Use of latest version of TOML parser