Refactor fetching of wathola receiver's delivery report using special batch Job (#4460)
* Reimplementing fetching of wathola report with K8s job

This change targets the problem of how to get report from cluster.
Clusters may have different networking setup, and it might not be
possible to directly make HTTP request from outside of cluster.

Previous approach used to guess an external address of cluster. That for
sure fails on OpenShift deployed on AWS.

This approach deploys a special Job that, being inside cluster, can
download a report and print it in it's logs. Then test client can fetch
logs of completed job, and parse it, replay the logs, and process report
further.

* Removal of unneeded external node address package

* Fixing lints & boilerplate

* spec.template.spec.restartPolicy=never

* Apply @devguyio suggestions for test/upgrade/README.md

Co-authored-by: Ahmed Abdalla Abdelrehim <aabdelre@redhat.com>

* Changes after review

Co-authored-by: Ahmed Abdalla Abdelrehim <aabdelre@redhat.com>