Bump golang linter version and timeout (#4808)
* [#4796] Bump golang linter to latest version

Bump golangci-lint to v1.35

Signed-off-by: Sameer Vohra <vsameer@vmware.com>

* [#4796] Bump golang lint timeout to 10m

Looking at the history of the "Code Style" workflow on eventing, there
are a bunch of failures with the lint task taking > 5mins and therefore
timing out. Bump the timeout to 10m to prevent the timeout. Successful
runs also end up taking ~4-4.5 mins so we're right on the threshold.

Signed-off-by: Sameer Vohra <vsameer@vmware.com>