Fix PingSource adapter metrics (#6615)
The PingSource was not exporting metrics on the advertised port
on the Deployment (metrics, 9090) but it was using 9092 (the 
eventing default).

In addition, any component using the adapter/v2 code suffers
from unexported metrics because the exporter was created
only after an update to `config-observability` since
the watcher was started before the observer was registered as
side effect of calling `CreateLogger`.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

Fixes #

<!-- Please include the 'why' behind your changes if no issue exists -->

## Proposed Changes

<!-- Please categorize your changes:
- :gift: Add new feature
- :bug: Fix bug
- :broom: Update or clean up current behavior
- :wastebasket: Remove feature or internal logic
-->

- Fix PingSource adapter metrics

### Pre-review Checklist

<!-- If these boxes are not checked, you will be asked to complete these
requirements or explain why they do not apply to your PR. -->

- [ ] **At least 80% unit test coverage**
- [ ] **E2E tests** for any new behavior
- [ ] **Docs PR** for any user-facing impact
- [ ] **Spec PR** for any new API feature
- [ ] **Conformance test** for any change to the spec

**Release Note**

<!--
:page_facing_up: If this change has user-visible impact, write a release
note in the block
below. Include the string "action required" if additional action is
required of
users switching to the new release, for example in case of a breaking
change.

Write as if you are speaking to users, not other Knative contributors.
If this
change has no user-visible impact, no release note is needed.
-->

```release-note
Fix PingSource adapter metrics
```

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>