Revert "Change subscription patch logic to ensure resource version (#6670) (#6724)
This reverts commit 4d6e1fc98d1a98c304566fa19843bcbb3fdea658.

It has the side effect of dropping channel spec fields, so even
immutable
fields are dropped, hence channels will fail to get updated.

We need to re-evalute the approach to fix the orginal issue:
https://github.com/knative/eventing/issues/6636, patch will always
have edge cases that will lead the original bug because subscriptions
are reconciled independently from each other (and potentially by 
multiple controller replicas), so update is the only way of having
concurrency control at the resource level but we should make sure
that we're preserving unknown fields when updating channelables.