EventType v1b2 API addition (#6893)
Fixes #6887

Following the pattern from
[here](https://github.com/knative/eventing/pull/2466), by just adding
the types

<!-- Please include the 'why' behind your changes if no issue exists -->

## Proposed Changes

<!-- Please categorize your changes:
- :gift: Add new feature
- :bug: Fix bug
- :broom: Update or clean up current behavior
- :wastebasket: Remove feature or internal logic
-->

- adding just `v1beta2` api for `EventType`

**Release Note**

<!--
:page_facing_up: If this change has user-visible impact, write a release
note in the block
below. Include the string "action required" if additional action is
required of
users switching to the new release, for example in case of a breaking
change.

Write as if you are speaking to users, not other Knative contributors.
If this
change has no user-visible impact, no release note is needed.
-->

```release-note
```

Signed-off-by: Matthias Wessendorf <mwessend@redhat.com>